chore(refactor): tidied up icons and general components.
