+++ src/lib/general/modal.svelte
@@ -1,8 +1,7 @@
 <script>
 	export let title,
 		open,
-		lg = false,
-		component;
+		lg = false;
 </script>
 
 {#if $open}
@@ -15,9 +14,6 @@
 			on:click={() => open.set(false)}>×</button
 		>
 		<h1 class="text-2xl font-bold">{title}</h1>
-		{#if component}
-			<svelte:component this={component} />
-		{/if}
 		<slot />
 	</div>
 {/if}
